Update to r2dbc-postgres 1.0.0 and r2dbc-pool 1.0.0

+++ core/src/main/scala/akka/persistence/r2dbc/internal/R2dbcExecutor.scala
@@ -41,27 +41,27 @@ import reactor.core.publisher.Mono
     }
   }
 
-  def updateOneInTx(stmt: Statement)(implicit ec: ExecutionContext): Future[Int] =
+  def updateOneInTx(stmt: Statement)(implicit ec: ExecutionContext): Future[Long] =
     stmt.execute().asFuture().flatMap { result =>
-      result.getRowsUpdated.asFuture().map(_.intValue())(ExecutionContexts.parasitic)
+      result.getRowsUpdated.asFuture().map(_.longValue())(ExecutionContexts.parasitic)
     }
 
-  def updateBatchInTx(stmt: Statement)(implicit ec: ExecutionContext): Future[Int] = {
-    val consumer: BiConsumer[Int, Integer] = (acc, elem) => acc + elem.intValue()
+  def updateBatchInTx(stmt: Statement)(implicit ec: ExecutionContext): Future[Long] = {
+    val consumer: BiConsumer[Long, java.lang.Long] = (acc, elem) => acc + elem.longValue()
     Flux
       .from[Result](stmt.execute())
       .concatMap(_.getRowsUpdated)
-      .collect(() => 0, consumer)
+      .collect(() => 0L, consumer)
       .asFuture()
   }
 
   def updateInTx(statements: immutable.IndexedSeq[Statement])(implicit
-      ec: ExecutionContext): Future[immutable.IndexedSeq[Int]] =
+      ec: ExecutionContext): Future[immutable.IndexedSeq[Long]] =
     // connection not intended for concurrent calls, make sure statements are executed one at a time
-    statements.foldLeft(Future.successful(Vector.empty[Int])) { (acc, stmt) =>
+    statements.foldLeft(Future.successful(Vector.empty[Long])) { (acc, stmt) =>
       acc.flatMap { seq =>
         stmt.execute().asFuture().flatMap { res =>
-          res.getRowsUpdated.asFuture().map(seq :+ _.intValue())(ExecutionContexts.parasitic)
+          res.getRowsUpdated.asFuture().map(seq :+ _.longValue())(ExecutionContexts.parasitic)
         }
       }
     }
@@ -147,7 +147,7 @@ class R2dbcExecutor(val connectionFactory: ConnectionFactory, log: Logger, logDb
   /**
    * One update statement with auto commit.
    */
-  def updateOne(logPrefix: String)(statementFactory: Connection => Statement): Future[Int] =
+  def updateOne(logPrefix: String)(statementFactory: Connection => Statement): Future[Long] =
     withAutoCommitConnection(logPrefix) { connection =>
       updateOneInTx(statementFactory(connection))
     }
@@ -155,7 +155,7 @@ class R2dbcExecutor(val connectionFactory: ConnectionFactory, log: Logger, logDb
   /**
    * Update statement that is constructed by several statements combined with `add()`.
    */
-  def updateInBatch(logPrefix: String)(statementFactory: Connection => Statement): Future[Int] =
+  def updateInBatch(logPrefix: String)(statementFactory: Connection => Statement): Future[Long] =
     withConnection(logPrefix) { connection =>
       updateBatchInTx(statementFactory(connection))
     }
@@ -164,7 +164,7 @@ class R2dbcExecutor(val connectionFactory: ConnectionFactory, log: Logger, logDb
    * Several update statements in the same transaction.
    */
   def update(logPrefix: String)(
-      statementsFactory: Connection => immutable.IndexedSeq[Statement]): Future[immutable.IndexedSeq[Int]] =
+      statementsFactory: Connection => immutable.IndexedSeq[Statement]): Future[immutable.IndexedSeq[Long]] =
     withConnection(logPrefix) { connection =>
       updateInTx(statementsFactory(connection))
     }
